src/python/requirements.txt
mysqlclient
mysql-connector-python
python-dotenv
pytest
pytest-cov
dotenv
